+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+## Unreleased 
+- Adds `Overstrike.new`.
 ## 0.1.337 (Aug 4, 2014)
 - Adds `Overstrike.assign`.
 - Adds `Overstrike.isEmpty`.