+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased 
 - Adds `Overstrike.new`.
+- Adds `Overstrike.noop`.
+
 ## 0.1.337 (Aug 4, 2014)
 - Adds `Overstrike.assign`.
 - Adds `Overstrike.isEmpty`.